requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
fastapi 
uvicorn

torch==1.12.1
torchvision==0.13.1
gitpython
ipython
shapely>=2.0.0
sahi>=0.11.9
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
psutil
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1
tqdm>=4.64.0
timm==0.6.11
tensorboard>=2.4.1
pandas==1.3.5
seaborn>=0.11.0
omegaconf>=2.2.3
yolov5>=7.0.6
ultralytics==8.0.147
pytorch-metric-learning==1.6.3
